tools: build sub-commands
